refactor: extract bot inactivity comment builders to bot-inactivity-comments.js
Resolves #1598

The four comment builder functions (buildWarningComment, buildClosureComment,
buildLinkedPRClosedComment, buildBlockedCheckinComment) and the two HTML
marker constants (WARN_MARKER, BLOCKED_CHECKIN_MARKER) have been moved from
bot-inactivity.js into a new sibling file bot-inactivity-comments.js.

This follows the established *.js <-> *-comments.js pattern used by the
commands/ bots (assign, finalize, unassign), making the inactivity bot
consistent with the rest of the bot system.

Changes:
- Create .github/scripts/bot-inactivity-comments.js with SPDX header,
  the two HTML markers, and all four builder functions exported via
  module.exports.
- Update buildWarningComment and buildClosureComment signatures to accept
  duration values (warnAfterMs, closeAfterMs) as parameters instead of
  reading module-level constants, keeping WARN_AFTER_MS, CLOSE_AFTER_MS,
  and BLOCKED_CHECKIN_AFTER_MS as a single source of truth in
  bot-inactivity.js.
- Replace inline definitions in bot-inactivity.js with a
  require('./bot-inactivity-comments') import.
- Update call sites in handleStaleItem to pass the duration constants into
  the builders.

All 34 existing tests in tests/test-inactivity-bot.js continue to pass
without modification. ESLint is clean. No behavioral change.
